Preparation for score(board) freeze

+++ html/includes/action.php
@@ -253,36 +253,42 @@
         }
     }
     else if ($_POST["action"] === "momentum") {
-        $done = false;
-        $last_update = fetchScalar("SELECT last_update()");
-        $result = fetchAll("SELECT UNIX_TIMESTAMP(ts) AS ts, value FROM cache WHERE name=:name", array("name" => Cache::MOMENTUM));
-
-        if ($result) {
-            // Reference: https://stackoverflow.com/a/15273676
-            ignore_user_abort(true);
-            set_time_limit(0);
-
-            ob_start();
-            echo $result[0]["value"];
-            header('Connection: close');
-            header('Content-Length: ' . ob_get_length());
-            header("Content-Encoding: none");  // Dirty "patch" for bypassing encoders stall
-            ob_end_flush();
-            ob_flush();
-            flush();
-
-            $done = true;
+        if (!isAdmin() && file_exists("momentum.json")) {
+            $result = file_get_contents("momentum.json");
+            echo $result;
         }
+        else {
+            $done = false;
+            $last_update = fetchScalar("SELECT last_update()");
+            $result = fetchAll("SELECT UNIX_TIMESTAMP(ts) AS ts, value FROM cache WHERE name=:name", array("name" => Cache::MOMENTUM));
 
-        if (!$done || ($result[0]["ts"] != $last_update)) {
-            $result = json_encode(getMomentum());
-            if ($last_update === fetchScalar("SELECT last_update()")) {  // Note: safety check to prevent dirty-write
-                execute("DELETE FROM cache WHERE name=:name", array("name" => Cache::MOMENTUM));
-                execute("INSERT INTO cache(name, value, ts) VALUES(:name, :value, FROM_UNIXTIME(:last_update))", array("name" => Cache::MOMENTUM, "value" => $result, "last_update" => $last_update));
+            if ($result) {
+                // Reference: https://stackoverflow.com/a/15273676
+                ignore_user_abort(true);
+                set_time_limit(0);
+
+                ob_start();
+                echo $result[0]["value"];
+                header('Connection: close');
+                header('Content-Length: ' . ob_get_length());
+                header("Content-Encoding: none");  // Dirty "patch" for bypassing encoders stall
+                ob_end_flush();
+                ob_flush();
+                flush();
+
+                $done = true;
             }
 
-            if (!$done)
-                echo $result;
+            if (!$done || ($result[0]["ts"] != $last_update)) {
+                $result = json_encode(getMomentum());
+                if ($last_update === fetchScalar("SELECT last_update()")) {  // Note: safety check to prevent dirty-write
+                    execute("DELETE FROM cache WHERE name=:name", array("name" => Cache::MOMENTUM));
+                    execute("INSERT INTO cache(name, value, ts) VALUES(:name, :value, FROM_UNIXTIME(:last_update))", array("name" => Cache::MOMENTUM, "value" => $result, "last_update" => $last_update));
+                }
+
+                if (!$done)
+                    echo $result;
+            }
         }
     }
 //     else if ($_POST["action"] === "momentum") {
